Moved view and gene switching functionality to main views folder for use in all views. Now includes option to switch only the gene, only the view, or both. Added tsdocs
